add output log and optimize the cleanup process

+++ del_animations.py
@@ -15,12 +15,14 @@ def read_unused_animations(path):
 def delete_animations():
     global unused_animations
     size = 0
-    for filename in unused_animations:
-        filename = filename.strip()
-        # print filename
-        size = size + os.path.getsize(filename)
-        os.remove(filename)
-
-    print "delete animations size=" + str(size / 1024) + "K"
-    return size
+    if len(unused_animations) == 0:
+        print 'no animations to be deleted'
+    else:
+        for filename in unused_animations:
+            filename = filename.strip()
+            # print filename
+            size = size + os.path.getsize(filename)
+            os.remove(filename)
+            print "deleted animations size=" + str(size / 1024) + "K"
+    return size, unused_animations
