+++ del_drawables.py
@@ -15,12 +15,15 @@ def read_unused_styles(path):
 def delete_pics():
     global unused_pics
     size = 0
-    for filename in unused_pics:
-        # path = project_path + os.sep + filename
-        print filename
-        size += os.path.getsize(filename)
-        os.remove(filename)
-    print "delete drawables size=" + str(size / 1024) + "K"
-    return size
+    if len(unused_pics) == 0:
+        print 'no drawables to be deleted'
+    else:
+        for filename in unused_pics:
+            # path = project_path + os.sep + filename
+            # print filename
+            size += os.path.getsize(filename)
+            os.remove(filename)
+        print "deleted drawables size=" + str(size / 1024) + "K"
+    return size, unused_pics
