+++ del_layouts.py
@@ -15,13 +15,16 @@ def read_unused_layouts(path):
 def delete_layouts():
     global unused_layouts
     size = 0
-    for filename in unused_layouts:
-        filename = filename.strip()
-        # print filename
-        size = size + os.path.getsize(filename)
-        os.remove(filename)
+    if len(unused_layouts) == 0:
+        print 'no layouts to be deleted'
+    else:
+        for filename in unused_layouts:
+            filename = filename.strip()
+            # print filename
+            size = size + os.path.getsize(filename)
+            os.remove(filename)
 
-    print "delete layouts size=" + str(size / 1024) + "K"
-    return size
+        print "deleted layouts size=" + str(size / 1024) + "K"
+    return size, unused_layouts
